# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Insurance Pricing API",
    "summary": "HTTP inference API for the insurance premium pricing model.",
    "description": "Production-style FastAPI surface for serving insurance pricing predictions.\n\n## Scope\n- score raw insurance business records\n- expose frequency, severity, and final premium predictions\n- keep the served model bundle pinned at startup through `INSURANCE_PRICING_RUN_ID`\n\n## Request Contract\nSend the raw columns from the historical scoring dataset, such as the ones present in `data/test.csv`. Feature engineering stays inside the service layer, so clients only need to send business-facing fields.\n\n## Response Headers\nEvery response includes `X-Request-ID` and `X-Process-Time-Ms` headers to help trace calls across clients, logs, and deployment environments.\n\n## Batch Behavior\nBatch endpoints preserve input order and return the optional `index` field when it is provided by the caller.\n\n## Product Quote Endpoints\nThe public `/quotes` endpoints persist end-user quote history scoped to an anonymous `X-Client-ID` header so the Next.js frontend can retrieve prior simulations.",
    "contact": {
      "name": "Insurance Pricing Team"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a customer account",
        "operationId": "register_account",
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentialsInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "409": {
            "description": "A user already exists with this email."
          },
          "503": {
            "description": "User persistence is unavailable."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate a user account",
        "operationId": "login_account",
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentialsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials."
          },
          "503": {
            "description": "Authentication is unavailable."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Inspect the current authenticated session",
        "operationId": "get_auth_session",
        "parameters": [
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/quotes": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Create a persisted quote",
        "description": "Scores one insurance record, persists the input payload and the resulting quote, and returns the stored quote detail for the current customer context.",
        "operationId": "create_quote",
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionInput",
                "description": "Raw business-facing fields for one insurance contract. The payload mirrors the technical prediction contract used by the model."
              },
              "examples": {
                "single_record": {
                  "summary": "Single contract quote example",
                  "value": {
                    "index": 50000,
                    "bonus": 0.58,
                    "type_contrat": "Maxi",
                    "duree_contrat": 1,
                    "anciennete_info": 1,
                    "freq_paiement": "Yearly",
                    "paiement": "No",
                    "utilisation": "Retired",
                    "code_postal": 28388,
                    "conducteur2": "No",
                    "age_conducteur1": 66,
                    "age_conducteur2": 0,
                    "sex_conducteur1": "F",
                    "sex_conducteur2": "",
                    "anciennete_permis1": 34,
                    "anciennete_permis2": 0,
                    "anciennete_vehicule": 16.0,
                    "cylindre_vehicule": 1239,
                    "din_vehicule": 55,
                    "essence_vehicule": "Gasoline",
                    "marque_vehicule": "RENAULT",
                    "modele_vehicule": "CLIO",
                    "debut_vente_vehicule": 16,
                    "fin_vente_vehicule": 15,
                    "vitesse_vehicule": 150,
                    "type_vehicule": "Tourism",
                    "prix_vehicule": 10321,
                    "poids_vehicule": 830
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Persisted quote detail for the current customer context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "The required X-Client-ID header is missing or invalid."
          },
          "503": {
            "description": "Quote persistence is unavailable because PostgreSQL is not ready."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "List persisted quotes",
        "description": "Returns the quote history associated with the current browser context or authenticated account, ordered from the most recent to the oldest quote.",
        "operationId": "list_quotes",
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted quote history for the current customer context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteListResponse"
                }
              }
            }
          },
          "400": {
            "description": "The required X-Client-ID header is missing or invalid."
          },
          "503": {
            "description": "Quote persistence is unavailable because PostgreSQL is not ready."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{quote_id}": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "Get a persisted quote",
        "description": "Returns one persisted quote when it belongs to the current browser context, the current account, or when it is accessed by an administrator.",
        "operationId": "get_quote",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted quote detail for the current customer context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "The required X-Client-ID header is missing or invalid."
          },
          "404": {
            "description": "The quote does not exist for the current customer context."
          },
          "503": {
            "description": "Quote persistence is unavailable because PostgreSQL is not ready."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List user accounts for administrators",
        "operationId": "admin_list_users",
        "parameters": [
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Deactivate a user account",
        "operationId": "admin_deactivate_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "User not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/quotes": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List recent quotes for administrators",
        "operationId": "admin_list_quotes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuoteListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/quotes/{quote_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete a persisted quote",
        "operationId": "admin_delete_quote",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Quote not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthCredentialsInput": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email used to create or access the account."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Plain-text password submitted by the user."
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "AuthCredentialsInput",
        "examples": [
          {
            "email": "client@nova-assurances.fr",
            "password": "mon-mot-de-passe"
          }
        ]
      },
      "AuthSessionResponse": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated",
            "description": "Whether a valid session is currently active."
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authenticated user summary when a session exists."
          },
          "session_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Token",
            "description": "Opaque session token returned to the trusted frontend BFF."
          },
          "expires_at_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At Utc",
            "description": "UTC timestamp at which the returned session expires."
          }
        },
        "type": "object",
        "required": [
          "authenticated"
        ],
        "title": "AuthSessionResponse",
        "examples": [
          {
            "authenticated": true,
            "expires_at_utc": "2026-04-22T13:00:00+00:00",
            "session_token": "opaque-session-token",
            "user": {
              "created_at_utc": "2026-03-23T13:00:00+00:00",
              "email": "client@nova-assurances.fr",
              "id": "7416b54b-26a0-4e79-9c27-aea0101f8d55",
              "is_active": true,
              "role": "customer"
            }
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}